refine code: create pub fn sign_multisig_txn_to_file for public using
